code/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vascular-reconstruction"
version = "0.1.0"
description = "Physics-Informed Gaussian Splatting for Vascular Reconstruction"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "numpy>=2.0.0",
    "pillow>=10.0.0",
    "trimesh>=4.0.0",
    "tqdm>=4.66.0",
    "kagglehub>=1.0.0",
    "rtree>=1.0.0",
    "scipy>=1.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/vascular_reconstruction"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "C4", "RET", "SIM", "PTH"]
ignore = []

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
